+++ src/app/(pages)/posts/[id]/components/NewsBody.tsx
@@ -10,6 +10,7 @@ import { CommentsCashType } from "@/types/CashTypes";
 import DOMPurify from "isomorphic-dompurify";
 import Image from "next/image";
 import Comment from "./Comment";
+import BookmarkButton from "./BookmarkButton";
 import ReadPost from "./ReadPost";
 import RenderTags from "./RenderTags";
 import SharePost from "./SharePost";
@@ -41,7 +42,10 @@ const NewsBody = ({ post }: { post: any }) => {
       />
       <div className="flex items-center justify-between gap-3 px-3">
         {/* <p className="text-center">{readingTimeEstimate.text}</p> */}
-        <ReadPost text={text} />
+        <div className="flex items-center gap-2">
+          <ReadPost text={text} />
+          <BookmarkButton postId={post._id} />
+        </div>
         <NexcaMark master={post.masterEditor} />
       </div>
       <div
